Refactoring for clarity; removed dead code
